refactor(navigation): split AppNavigator into focused modules
Extract transition specs to transitions.ts, Android sheet wrappers to
SheetRoutes.tsx, tab icons to TabIcons.tsx, and the animated tab button
to TabBarButton.tsx. AppNavigator drops from 823 to 450 lines and now
focuses on route declarations. No route names, params, or animation
behavior changed.
